re org to fix ci
