+++ web/api/helpers/app-store/graphql/get-app-rating.generated.ts
