+++ web/api/v2/minikit/app-metadata/[app_id]/index.ts
@@ -1,8 +1,8 @@
+import { formatAppMetadata } from "@/api/helpers/app-store";
 import { errorResponse } from "@/api/helpers/errors";
 import { getAPIServiceGraphqlClient } from "@/api/helpers/graphql";
 import { verifyHashedSecret } from "@/api/helpers/utils";
 import { parseLocale } from "@/lib/languages";
-import { formatAppMetadata } from "@/lib/utils";
 import { NextRequest, NextResponse } from "next/server";
 import { getSdk as fetchApiKeySdk } from "../../graphql/fetch-api-key.generated";
 import { getSdk as getAppMetadataSdk } from "./graphql/get-app-metadata.generated";
