+++ web/lib/utils.ts
@@ -1,23 +1,13 @@
-import { getAppRating } from "@/api/helpers/app-ratings/app-ratings";
 import { Role_Enum } from "@/graphql/graphql";
 import { Auth0EmailUser, Auth0User } from "@/legacy/lib/types";
 import { VerificationLevel } from "@worldcoin/idkit-core";
-import { getLocalisedCategory } from "./categories";
 import {
   ORB_SEQUENCER,
   ORB_SEQUENCER_STAGING,
   PHONE_SEQUENCER,
   PHONE_SEQUENCER_STAGING,
-  whitelistedAppsPermit2,
 } from "./constants";
-import { generateExternalNullifier } from "./hashing";
-import {
-  AppStatsReturnType,
-  AppStoreFormattedFields,
-  AppStoreMetadataDescription,
-  AppStoreMetadataFields,
-  Auth0SessionUser,
-} from "./types";
+import { Auth0SessionUser } from "./types";
 
 // Sequencer mapping
 export const sequencerMapping: Record<
@@ -199,101 +189,6 @@ export const tryParseJSON = (jsonString: string) => {
   }
 
   return null;
-};
-
-export const formatAppMetadata = async (
-  appData: AppStoreMetadataFields,
-  appStats: AppStatsReturnType,
-  locale: string = "en",
-): Promise<AppStoreFormattedFields> => {
-  const { app, ...appMetadata } = appData;
-  const appStat: number =
-    appStats.find((stat) => stat.app_id === appMetadata.app_id)?.unique_users ??
-    0;
-
-  const appRating = await getAppRating(appMetadata.app_id);
-
-  const localisedContent = appMetadata.localisations?.[0];
-
-  // We pick default description if localised content is not available
-  const description: AppStoreMetadataDescription = tryParseJSON(
-    localisedContent?.description ?? appMetadata.description,
-  );
-
-  const {
-    localisations,
-    is_reviewer_world_app_approved,
-    ...appMetadataWithoutLocalisations
-  } = appMetadata;
-
-  const name = localisedContent?.name ?? appMetadata.name;
-
-  // Check if the app is whitelisted for permit2
-  const permit2Tokens = whitelistedAppsPermit2.includes(appMetadata.app_id)
-    ? ["all"]
-    : appMetadata.permit2_tokens;
-
-  return {
-    ...appMetadataWithoutLocalisations,
-    name: name,
-    app_rating: appRating,
-    world_app_button_text:
-      localisedContent?.world_app_button_text ??
-      appMetadata.world_app_button_text,
-    world_app_description:
-      localisedContent?.world_app_description ??
-      appMetadata.world_app_description,
-    short_name: localisedContent?.short_name ?? appMetadata.short_name ?? name,
-    logo_img_url: getCDNImageUrl(
-      appMetadata.app_id,
-      appMetadata.logo_img_url,
-      appMetadata.verification_status === "verified",
-    ),
-    showcase_img_urls: appMetadata.showcase_img_urls?.map((url: string) =>
-      getCDNImageUrl(
-        appMetadata.app_id,
-        url,
-        appMetadata.verification_status === "verified",
-      ),
-    ),
-    hero_image_url: getCDNImageUrl(
-      appMetadata.app_id,
-      appMetadata.hero_image_url,
-      appMetadata.verification_status === "verified",
-    ),
-    description: {
-      overview: description?.description_overview ?? "",
-      how_it_works: description?.description_how_it_works ?? "",
-      how_to_connect: description?.description_connect ?? "",
-    },
-    ratings_external_nullifier: generateExternalNullifier(
-      `${appMetadata.app_id}_app_review`,
-    ).digest,
-    unique_users: appStat,
-    category: getLocalisedCategory(appMetadata.category, locale) ?? {
-      id: "other",
-      name: "Other",
-    },
-    show_in_app_store: is_reviewer_world_app_approved,
-    team_name: app.team.name ?? "",
-    permit2_tokens: permit2Tokens,
-  };
-};
-
-// Cached thus this is not that expensive
-export const rankApps = (
-  apps: AppStoreFormattedFields[],
-  appStats: AppStatsReturnType,
-) => {
-  return apps.sort((a, b) => {
-    const aStat = appStats.find((stat) => stat.app_id === a.app_id);
-    const bStat = appStats.find((stat) => stat.app_id === b.app_id);
-
-    return (
-      (bStat?.unique_users_last_7_days ?? 0) -
-      (aStat?.unique_users_last_7_days ?? 0)
-    );
-  });
 };
 
 // Helper function to ensure uploaded images are png or jpg. Otherwise hasura trigger will fail
